progress

+++ api/types.go
@@ -23,8 +23,8 @@ type PullRequest struct {
 }
 
 type PullProgress struct {
-	Total     int     `json:"total"`
-	Completed int     `json:"completed"`
+	Total     int64   `json:"total"`
+	Completed int64   `json:"completed"`
 	Percent   float64 `json:"percent"`
 }
